chore: restructure bin files
