+++ .gitignore
@@ -2,7 +2,7 @@
 node_modules
 /dist
 tsconfig.tsbuildinfo
-
+.eodash
 # local env files
 .env.local
 .env.*.local
