+++ bin/serverConfig.js
@@ -1,18 +1,18 @@
 #!/usr/bin/env node
 
-import { build, createServer, defineConfig, preview, searchForWorkspaceRoot } from "vite"
+import { defineConfig, searchForWorkspaceRoot } from "vite"
 // Plugins
 import vue from '@vitejs/plugin-vue';
 import vuetify, { transformAssetUrls } from 'vite-plugin-vuetify';
 
 // Utilities
 import { fileURLToPath, URL } from 'url';
-import { execPath, configPath, appPath, dotEodashPath } from "./utils.js";
-import { readFile, writeFile, rm, cp } from "fs/promises";
+import { configPath, appPath, dotEodashPath } from "./utils.js";
+import { readFile } from "fs/promises";
 import { update } from "./update.js";
 
 
-const indexHtml = `
+export const indexHtml = `
 <!DOCTYPE html>
 <html lang="en">
 
@@ -29,7 +29,7 @@ const indexHtml = `
 </body>
 </html>`
 
-const serverConfig = defineConfig(({ mode, command }) => {
+export const serverConfig = defineConfig(({ mode, command }) => {
   return {
     base: '',
     cacheDir: dotEodashPath + '/cache',
@@ -84,50 +84,6 @@ const serverConfig = defineConfig(({ mode, command }) => {
   }
 });
 
-
-
-export const createDevServer = async () => {
-  await update()
-  const server = await createServer(serverConfig({ mode: 'development', command: 'serve' }))
-  await server.listen()
-  server.printUrls()
-  server.bindCLIShortcuts({ print: true })
-}
-
-export const buildApp = async () => {
-  await update()
-  const htmlPath = appPath + 'index.html'
-  await writeFile(htmlPath, indexHtml).then(async () => {
-    await build(serverConfig({ mode: 'production', command: 'build' }))
-    await rm(htmlPath).catch(() => {
-      console.error('failed to remove index.html')
-    })
-  }).catch(() => {
-    console.error('failed to create production entry point')
-  })
-  if (appPath.includes('node_modules')) {
-    await cp(appPath + 'dist', execPath + '/app', { recursive: true }).then(() => {
-      console.info('dashboard built successfully')
-    }).catch((e) => {
-      console.error(e)
-    })
-  }
-}
-
-
-export async function previewApp() {
-  const previewServer = await preview({
-    root: execPath,
-    preview: {
-      port: 8080,
-      open: true,
-    },
-    build: {
-      outDir: 'app'
-    }
-  })
-  previewServer.printUrls()
-}
 
 
 /**
